Home Button Redesign
Completely rethought the way I designed the home button from the original 5 model. The shape is the same, but I've removed the square cutout and placed two different shapes behind the button itself. The button is expected to be textured in glass with the underlying materials being "projected" on the outer surface through the glass.